+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 ## [Unreleased]
 
 ### Added
@@ -9,11 +10,12 @@
 ### Changed
 
 - [TD-3957] Structure profiles are now validated and expanded
+- [TD-3952] Data structure types now support multiple metadata views
 
 ### Fixed
 
 - [TD-3959] Report `rule_result_created` when result is created
-- [TD-3908] Timeout on node retrieval 
+- [TD-3908] Timeout on node retrieval
 
 ## [4.25.0] 2021-07-26